refactor: update pruneClosedTool to use TaskTrellisService
